Move getLicenceInfo to match import structure
